pyproject.toml
[tool.poetry]
name = "ai-compliance-agent"
version = "0.1.0"
description = "AI-powered Compliance & Onboarding Agent for Fintech"
authors = ["Antigravity <antigravity@google.deepmind>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
redis = "^5.0.1"
arq = "^0.25.0"
playwright = "^1.41.0"
httpx = "^0.26.0"
beautifulsoup4 = "^4.12.3"
jinja2 = "^3.1.3"
tenacity = "^8.2.3"
# AI Clients (to be used via abstraction)
openai = "^1.10.0"
google-generativeai = "^0.3.2"
weasyprint = "^60.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
